Extract host (de)serialization
Created a new module app.serialization. Extracted all Host serialization
and deserialization logic into that module. Used the existing paradigm
from CanonicalFacts and Facts classes for the Host itself too. All the
logic and behavior remained unchanged.
